Made changes to UI and fixed the broken FE tests
