+++ app/src/features/projects/view/ProjectAttachments.tsx
@@ -6,18 +6,15 @@ import AttachmentsList from 'components/attachments/AttachmentsList';
 import FileUpload from 'components/attachments/FileUpload';
 import { IUploadHandler } from 'components/attachments/FileUploadItem';
 import ComponentDialog from 'components/dialog/ComponentDialog';
-import { H2ButtonToolbar } from 'components/toolbar/ActionToolbars';
+import { H3ButtonToolbar } from 'components/toolbar/ActionToolbars';
 import { useRestorationTrackerApi } from 'hooks/useRestorationTrackerApi';
-import {
-  IGetProjectAttachment,
-  IGetProjectForViewResponse,
-  IUploadAttachmentResponse
-} from 'interfaces/useProjectApi.interface';
-import React, { useCallback, useEffect, useState } from 'react';
+import { IGetProjectAttachment, IUploadAttachmentResponse } from 'interfaces/useProjectApi.interface';
+import React, { useState } from 'react';
 import { useParams } from 'react-router';
 
 export interface IProjectAttachmentsProps {
-  projectForViewData: IGetProjectForViewResponse;
+  attachmentsList: IGetProjectAttachment[];
+  getAttachments: (forceFetch: boolean) => void;
 }
 
 /**
@@ -25,38 +22,15 @@ export interface IProjectAttachmentsProps {
  *
  * @return {*}
  */
-const ProjectAttachments: React.FC<IProjectAttachmentsProps> = () => {
+const ProjectAttachments: React.FC<IProjectAttachmentsProps> = (props) => {
+  const { attachmentsList, getAttachments } = props;
   const urlParams = useParams();
   const projectId = urlParams['id'];
   const restorationTrackerApi = useRestorationTrackerApi();
 
   const [openUploadAttachments, setOpenUploadAttachments] = useState(false);
-  const [attachmentsList, setAttachmentsList] = useState<IGetProjectAttachment[]>([]);
 
-  const handleUploadAttachmentClick = () => {
-    setOpenUploadAttachments(true);
-  };
-
-  const getAttachments = useCallback(
-    async (forceFetch: boolean) => {
-      if (attachmentsList.length && !forceFetch) {
-        return;
-      }
-
-      try {
-        const response = await restorationTrackerApi.project.getProjectAttachments(projectId);
-
-        if (!response?.attachmentsList) {
-          return;
-        }
-
-        setAttachmentsList([...response.attachmentsList]);
-      } catch (error) {
-        return error;
-      }
-    },
-    [restorationTrackerApi.project, projectId, attachmentsList.length]
-  );
+  const handleUploadAttachmentClick = () => setOpenUploadAttachments(true);
 
   const getUploadHandler = (): IUploadHandler<IUploadAttachmentResponse> => {
     return (file, cancelToken, handleFileUploadProgress) => {
@@ -68,11 +42,6 @@ const ProjectAttachments: React.FC<IProjectAttachmentsProps> = () => {
       );
     };
   };
-
-  useEffect(() => {
-    getAttachments(false);
-    // eslint-disable-next-line
-  }, []);
 
   return (
     <>
@@ -86,12 +55,15 @@ const ProjectAttachments: React.FC<IProjectAttachmentsProps> = () => {
         <FileUpload uploadHandler={getUploadHandler()} />
       </ComponentDialog>
       <Paper>
-        <H2ButtonToolbar
+        <H3ButtonToolbar
           label="Documents"
           buttonLabel="Upload"
           buttonTitle="Upload Document"
           buttonStartIcon={<Icon path={mdiTrayArrowUp} size={1} />}
           buttonOnClick={handleUploadAttachmentClick}
+          buttonProps={{
+            variant: 'outlined'
+          }}
         />
         <Box px={3} pb={2}>
           <AttachmentsList projectId={projectId} attachmentsList={attachmentsList} getAttachments={getAttachments} />
