+++ app/src/pages/public/components/PublicAttachmentsList.tsx
@@ -1,9 +1,3 @@
-import Box from '@material-ui/core/Box';
-import Button from '@material-ui/core/Button';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogTitle from '@material-ui/core/DialogTitle';
 import Link from '@material-ui/core/Link';
 import { Theme } from '@material-ui/core/styles/createMuiTheme';
 import makeStyles from '@material-ui/core/styles/makeStyles';
@@ -14,15 +8,10 @@ import TableContainer from '@material-ui/core/TableContainer';
 import TableHead from '@material-ui/core/TableHead';
 import TablePagination from '@material-ui/core/TablePagination';
 import TableRow from '@material-ui/core/TableRow';
-import Typography from '@material-ui/core/Typography';
-import { mdiLockOpenVariantOutline, mdiLockOutline } from '@mdi/js';
-import Icon from '@mdi/react';
-import { DATE_FORMAT } from 'constants/dateTimeFormats';
-import { useRestorationTrackerApi } from 'hooks/useRestorationTrackerApi';
 import { IGetProjectAttachment } from 'interfaces/useProjectApi.interface';
 import React, { useState } from 'react';
 import { handleChangePage, handleChangeRowsPerPage } from 'utils/tablePaginationUtils';
-import { getFormattedDate, getFormattedFileSize } from 'utils/Utils';
+import { getFormattedFileSize } from 'utils/Utils';
 
 const useStyles = makeStyles((theme: Theme) => ({
   attachmentsTable: {
@@ -40,37 +29,12 @@ export interface IPublicAttachmentsListProps {
 
 const PublicAttachmentsList: React.FC<IPublicAttachmentsListProps> = (props) => {
   const classes = useStyles();
-  const [open, setOpen] = React.useState(false);
-  const restorationTrackerApi = useRestorationTrackerApi();
-  const preventDefault = (event: React.SyntheticEvent) => event.preventDefault();
 
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const [page, setPage] = useState(0);
 
-  const showRequestAccessDialog = () => {
-    setOpen(true);
-  };
+  const openAttachment = async (attachment: IGetProjectAttachment) => window.open(attachment.url);
 
-  const hideRequestAccessDialog = () => {
-    setOpen(false);
-  };
-
-  const openAttachment = async (attachment: IGetProjectAttachment) => {
-    try {
-      const response = await restorationTrackerApi.public.project.getAttachmentSignedURL(
-        props.projectId,
-        attachment.id
-      );
-
-      if (!response) {
-        return;
-      }
-
-      window.open(response);
-    } catch (error) {
-      return error;
-    }
-  };
   return (
     <>
       <TableContainer>
@@ -78,10 +42,7 @@ const PublicAttachmentsList: React.FC<IPublicAttachmentsListProps> = (props) => 
           <TableHead>
             <TableRow>
               <TableCell>Name</TableCell>
-              <TableCell>Last Modified</TableCell>
               <TableCell>File Size</TableCell>
-              <TableCell width="150px">Security Status</TableCell>
-              <TableCell></TableCell>
             </TableRow>
           </TableHead>
           <TableBody>
@@ -89,29 +50,11 @@ const PublicAttachmentsList: React.FC<IPublicAttachmentsListProps> = (props) => 
               props.attachmentsList.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((row, index) => (
                 <TableRow key={`${row.fileName}-${index}`}>
                   <TableCell scope="row">
-                    <Link
-                      underline="always"
-                      component="button"
-                      variant="body2"
-                      onClick={() => {
-                        if (row.securityToken) {
-                          showRequestAccessDialog();
-                        } else {
-                          openAttachment(row);
-                        }
-                      }}>
+                    <Link underline="always" component="button" variant="body2" onClick={() => openAttachment(row)}>
                       {row.fileName}
                     </Link>
                   </TableCell>
-                  <TableCell>{getFormattedDate(DATE_FORMAT.ShortDateFormatMonthFirst, row.lastModified)}</TableCell>
                   <TableCell>{getFormattedFileSize(row.size)}</TableCell>
-                  <TableCell>
-                    <Box display="flex" alignItems="center">
-                      <Icon path={row.securityToken ? mdiLockOutline : mdiLockOpenVariantOutline} size={1} />
-                      <Box ml={0.5}>{row.securityToken ? 'Secured' : 'Unsecured'}</Box>
-                    </Box>
-                  </TableCell>
-                  <TableCell></TableCell>
                 </TableRow>
               ))}
             {!props.attachmentsList.length && (
@@ -137,27 +80,6 @@ const PublicAttachmentsList: React.FC<IPublicAttachmentsListProps> = (props) => 
           }
         />
       )}
-
-      <Dialog open={open}>
-        <DialogTitle>Access Denied</DialogTitle>
-        <DialogContent>
-          <Typography variant="body1" color="textSecondary">
-            To access secure documents, please submit your request to{' '}
-            <Link
-              href="mailto:retoration-tracker@gov.bc.ca?subject=Restoration Tracker - Secure Document Access Request"
-              underline="always"
-              onClick={preventDefault}>
-              restoation-tracker@gov.bc.ca
-            </Link>
-            . A data manager will review your request and contact you as soon as possible.
-          </Typography>
-        </DialogContent>
-        <DialogActions>
-          <Button variant="contained" color="primary" onClick={hideRequestAccessDialog}>
-            OK
-          </Button>
-        </DialogActions>
-      </Dialog>
     </>
   );
 };
